move price code to pricer/main
