+++ apps/daemon/src/main.rs
@@ -10,9 +10,6 @@ use crate::version_updater::Client as VersionClient;
 use api_connector::AssetsClient;
 use fiat::FiatProviderFactory;
 use fiat_assets_updater::FiatAssetsUpdater;
-use pricer::client::PriceClient;
-use pricer::coingecko::CoinGeckoClient;
-use pricer::price_updater::PriceUpdater;
 use std::thread;
 use std::time::Duration;
 
@@ -20,15 +17,6 @@ use std::time::Duration;
 pub async fn main() {
     println!("daemon init");
     let settings = settings::Settings::new().unwrap();
-    let price_client = PriceClient::new(
-        &settings.redis.url,
-        &settings.postgres.url,
-        &settings.clickhouse.url,
-    )
-    .await
-    .unwrap();
-    let coingecko_client = CoinGeckoClient::new(settings.coingecko.key.secret);
-    let mut price_updater = PriceUpdater::new(price_client, coingecko_client);
     let mut version_client = VersionClient::new(&settings.postgres.url);
     let mut device_updater = DeviceUpdater::new(&settings.postgres.url);
     let assets_client = AssetsClient::new(settings.assets.url);
@@ -45,28 +33,6 @@ pub async fn main() {
         }
         Err(err) => {
             println!("update fiat assets error: {}", err)
-        }
-    }
-
-    // update assets
-    let result = price_updater.update_assets().await;
-    match result {
-        Ok(count) => {
-            println!("update assets: {}", count)
-        }
-        Err(err) => {
-            println!("update assets error: {}", err)
-        }
-    }
-
-    // update rates
-    let result = price_updater.update_fiat_rates().await;
-    match result {
-        Ok(count) => {
-            println!("update rates: {}", count)
-        }
-        Err(err) => {
-            println!("update rates error: {}", err)
         }
     }
 
@@ -92,40 +58,7 @@ pub async fn main() {
         }
     }
 
-    // updates charts
-    // let result = price_updater.update_charts().await;
-    // match result {
-    //     Ok(count) => {
-    //         println!("update charts: {}", count)
-    //     }
-    //     Err(err) => {
-    //         println!("update charts error: {}", err)
-    //     }
-    // }
-
     loop {
-        // updates prices
-        let result = price_updater.update_prices().await;
-        match result {
-            Ok(count) => {
-                println!("update prices: {}", count)
-            }
-            Err(err) => {
-                println!("update prices error: {}", err)
-            }
-        }
-
-        // update cache
-        let result = price_updater.update_cache().await;
-        match result {
-            Ok(count) => {
-                println!("update prices cache: {}", count)
-            }
-            Err(err) => {
-                println!("update prices cache error: {}", err)
-            }
-        }
-
         let result = tokenlist_client.update().await;
         match result {
             Ok(count) => {
