+++ gemstone/Cargo.toml
@@ -20,7 +20,7 @@ once_cell = "1.19.0"
 
 # Sui
 bcs = "0.1.6"
-sui-types = { git = "https://github.com/mystenlabs/sui", package = "sui-types", rev = "f5b55da420b2ed2c8bea46ff7ab62709d49f9c2e" }
+sui-types = { git = "https://github.com/mystenlabs/sui", package = "sui-types", tag = "mainnet-v1.19.1" }
 indexmap = { workspace = true }
 blake2 = { workspace = true }
 base64 = { workspace = true }
